+++ src/VirtoCommerce.ProfileExperienceApiModule.Data/Queries/AddressesQuery/BaseAddressesQueryBuilder.cs
@@ -1,0 +1,51 @@
+using System.Threading.Tasks;
+using GraphQL;
+using GraphQL.Types;
+using GraphQL.Types.Relay;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using VirtoCommerce.ProfileExperienceApiModule.Data.Models;
+using VirtoCommerce.ProfileExperienceApiModule.Data.Schemas;
+using VirtoCommerce.ProfileExperienceApiModule.Data.Services;
+using VirtoCommerce.Xapi.Core.BaseQueries;
+using VirtoCommerce.Xapi.Core.Helpers;
+
+namespace VirtoCommerce.ProfileExperienceApiModule.Data.Queries.AddressesQuery;
+
+public abstract class BaseAddressesQueryBuilder<TQuery> : SearchQueryBuilder<TQuery, MemberAddressSearchResult, MemberAddress, MemberAddressType>
+    where TQuery : BaseAddressesQuery
+{
+    private readonly IProfileAuthorizationService _profileAuthorizationService;
+
+    protected BaseAddressesQueryBuilder(IMediator mediator, IAuthorizationService authorizationService, IProfileAuthorizationService profileAuthorizationService)
+        : base(mediator, authorizationService)
+    {
+        _profileAuthorizationService = profileAuthorizationService;
+    }
+
+    protected override FieldType GetFieldType()
+    {
+        var builder = GraphTypeExtensionHelper
+            .CreateConnection<MemberAddressType, EdgeType<MemberAddressType>, MemberAddressConnectionType<MemberAddressType>, object>(Name)
+            .PageSize(DefaultPageSize);
+
+        ConfigureArguments(builder.FieldType);
+
+        builder.ResolveAsync(async context =>
+        {
+            var (query, response) = await Resolve(context);
+            return new MemberAddressConnection<MemberAddress>(response.Results, query.Skip, query.Take, response.TotalCount)
+            {
+                Facets = response.Facets,
+            };
+        });
+
+        return builder.FieldType;
+    }
+
+    protected override async Task BeforeMediatorSend(IResolveFieldContext<object> context, TQuery request)
+    {
+        await base.BeforeMediatorSend(context, request);
+        await _profileAuthorizationService.CheckAuthAsync(context, request);
+    }
+}
