+++ src/VirtoCommerce.ProfileExperienceApiModule.Data/Queries/AddressesQuery/CurrentOrganizationAddressesQueryBuilder.cs
@@ -1,54 +1,15 @@
-using System.Threading.Tasks;
-using GraphQL;
-using GraphQL.Types;
-using GraphQL.Types.Relay;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
-using VirtoCommerce.ProfileExperienceApiModule.Data.Models;
-using VirtoCommerce.ProfileExperienceApiModule.Data.Schemas;
 using VirtoCommerce.ProfileExperienceApiModule.Data.Services;
-using VirtoCommerce.Xapi.Core.BaseQueries;
-using VirtoCommerce.Xapi.Core.Helpers;
 
 namespace VirtoCommerce.ProfileExperienceApiModule.Data.Queries.AddressesQuery;
 
-public class CurrentOrganizationAddressesQueryBuilder : SearchQueryBuilder<CurrentOrganizationAddressesQuery, MemberAddressSearchResult, MemberAddress, MemberAddressType>
+public class CurrentOrganizationAddressesQueryBuilder : BaseAddressesQueryBuilder<CurrentOrganizationAddressesQuery>
 {
     protected override string Name => "currentOrganizationAddresses";
 
-    private readonly IProfileAuthorizationService _profileAuthorizationService;
-
     public CurrentOrganizationAddressesQueryBuilder(IMediator mediator, IAuthorizationService authorizationService, IProfileAuthorizationService profileAuthorizationService)
-        : base(mediator, authorizationService)
+        : base(mediator, authorizationService, profileAuthorizationService)
     {
-        _profileAuthorizationService = profileAuthorizationService;
-    }
-
-    protected override async Task BeforeMediatorSend(IResolveFieldContext<object> context, CurrentOrganizationAddressesQuery request)
-    {
-        await base.BeforeMediatorSend(context, request);
-        await _profileAuthorizationService.CheckAuthAsync(context, request);
-    }
-
-    protected override FieldType GetFieldType()
-    {
-        var builder = GraphTypeExtensionHelper
-            .CreateConnection<MemberAddressType, EdgeType<MemberAddressType>, MemberAddressConnectionType<MemberAddressType>, object>(Name)
-            .PageSize(DefaultPageSize);
-
-        ConfigureArguments(builder.FieldType);
-
-        builder.ResolveAsync(async context =>
-        {
-            var (query, response) = await Resolve(context);
-            return new MemberAddressConnection<MemberAddress>(response.Results, query.Skip, query.Take, response.TotalCount)
-            {
-                Facets = response.Facets,
-            };
-        });
-
-        return builder.FieldType;
     }
 }
-
-
